+++ .eslintrc
@@ -1,10 +1,9 @@
 {
   "env": {
-    "browser": true,
+    "node": true,
     "es6": true
   },
   "globals": {
-    "MashupPlatform": false,
     "Promise": false,
     "WeakMap": false
   },
