+++ package.json
@@ -1,32 +1,63 @@
 {
   "description": "This package.json file is only used for installing npm dependencies. But this is not an installable node package, but a WireCloud widget. Take a look into src/config.xml for more details about this widget",
   "devDependencies": {
+    "@babel/core": "^7.8.4",
+    "@babel/plugin-transform-runtime": "^7.8.3",
+    "@babel/polyfill": "^7.8.3",
+    "@babel/preset-env": "^7.8.4",
+    "@babel/runtime": "^7.8.4",
+    "babel-eslint": "^10.0.3",
+    "babel-loader": "^8.0.6",
+    "babel-plugin-add-module-exports": "^1.0.2",
+    "babel-plugin-istanbul": "^6.0.0",
+    "babel-preset-stage-0": "^6.24.1",
+    "babelify": "^10.0.0",
+    "browserify": "^16.5.0",
+    "browserify-babel-istanbul": "^0.4.0",
+    "browserify-istanbul": "^3.0.1",
+    "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
     "grunt": "^1.0.4",
-    "grunt-contrib-clean": "~1.1.0",
+    "grunt-babel": "^8.0.0",
+    "grunt-contrib-clean": "^2.0.0",
     "grunt-contrib-compress": "^1.6.0",
     "grunt-contrib-copy": "^1.0.0",
     "grunt-coveralls": "^2.0.0",
-    "grunt-curl": "^2.5.1",
-    "grunt-karma": "^3.0.1",
+    "grunt-eslint": "^23.0.0",
+    "grunt-karma": "^3.0.2",
+    "grunt-run": "^0.8.1",
     "grunt-strip-code": "^1.0.6",
     "grunt-text-replace": "~0.4.0",
-    "grunt-wirecloud": "^0.9.6",
+    "grunt-webpack": "^3.1.3",
+    "grunt-wirecloud": "^0.9.7",
     "gruntify-eslint": "^5.0.0",
-    "jasmine-core": "^2.99.1",
-    "karma": "^6.3.2",
-    "karma-chrome-launcher": "^1.0.1",
+    "jasmine-core": "^3.5.0",
+    "karma": "^6.2.0",
+    "karma-babel-preprocessor": "^8.0.1",
+    "karma-browserify": "^7.0.0",
+    "karma-chai": "^0.1.0",
+    "karma-chrome-launcher": "^3.1.0",
     "karma-coverage": "^2.0.1",
-    "karma-firefox-launcher": "^1.0.0",
-    "karma-jasmine": "^1.1.2",
-    "karma-junit-reporter": "^1.2.0",
-    "mock-applicationmashup": "^1.0.0-e",
+    "karma-firefox-launcher": "^1.2.0",
+    "karma-jasmine": "^2.0.1",
+    "karma-junit-reporter": "^2.0.1",
+    "karma-mocha": "^2.0.1",
+    "karma-mocha-reporter": "^2.2.5",
+    "mocha": "^7.0.1",
+    "sinon": "^8.1.1",
+    "watchify": "^3.11.1",
+    "webpack": "^4.41.5",
     "wirecloud-config-parser": "^0.2.3"
   },
   "private": true,
   "dependencies": {
+    "@turf/helpers": "^6.3.0",
     "@turf/turf": "^6.3.0",
+    "css-loader": "^5.2.4",
     "maplibre-gl": "^1.14.0",
+    "style-loader": "^2.0.0",
     "three": "^0.116.0",
+    "three-gltf-loader": "^1.111.0",
     "topojson-client": "^3.1.0",
     "topojson-server": "^3.0.1",
     "topojson-simplify": "^3.0.3"
