+++ src/index.html
@@ -3,12 +3,9 @@
     <head>
         <meta charset="utf-8" />
         <link rel="stylesheet" type="text/css" href="css/styles.css" />
-        <script src="lib/maplibre-gl/maplibre-gl.js"></script>
-        <link href="lib/maplibre-gl/maplibre-gl.css" rel="stylesheet" />
         <script src="lib/js/three.min.js"></script>
-        <script src="lib/js/turf.min.js"></script>
         <script src="lib/js/GLTFLoader.js"></script>
-        <script src="lib/mapbox-3dtiles/Mapbox3DTiles.js"></script>
+        <script src="lib/js/Mapbox3DTiles.js"></script>
         <title>MapLibre GL widget</title>
     </head>
     <body>
@@ -23,7 +20,6 @@
                 </div>
             </div>
         </div>
-        <script type="text/javascript" src="js/MapLibre.js"></script>
         <script type="text/javascript" src="js/main.js"></script>
     </body>
 </html>
